# swagger.yaml
openapi: 3.0.3
info:
  title: REST API - Gestion de produits, catégories et entreprises
  description: |
    API REST complète pour la gestion d'un système e-commerce avec :
    - Gestion des produits (simples et variables)
    - Gestion des catégories et sous-catégories
    - Gestion des entreprises
    - Système de commandes
    - Authentification et autorisation basée sur les rôles
    - Notifications système
    - Upload et optimisation d'images via Cloudinary
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: ISC
servers:
  - url: http://localhost:5000
    description: Serveur de développement
  - url: https://api.yourdomain.com
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /api/auth/login
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "password123"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT d'accès
        refreshToken:
          type: string
          description: Token de rafraîchissement
        expiresIn:
          type: integer
          description: Durée de vie du token en millisecondes
        username:
          type: string
        role:
          type: string
          enum: [superAdmin, productManager, orderManager, contentEditor]
        firstName:
          type: string
        lastName:
          type: string
          
    # Schémas des produits
...
    Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          example: "Vêtements"
        description:
          type: string
        image:
          type: string
          description: URL de l'image
        createdAt:
          type: string
          format: date-time
          
    # Schémas des entreprises
...
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
  responses:
    UnauthorizedError:
      description: Token d'authentification manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Connexion administrateur
      description: Authentifie un administrateur et retourne un token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/verify:
    get:
      tags:
        - Auth
      summary: Vérifier l'authentification
      description: Vérifie la validité du token et retourne les infos utilisateur
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token valide
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAuthenticated:
                    type: boolean
                  admin:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Routes des produits
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Lister les catégories
      description: Récupère toutes les catégories
      responses:
        '200':
          description: Liste des catégories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
                  
    post:
      tags:
        - Categories
      summary: Créer une catégorie
      description: Crée une nouvelle catégorie (permission requise)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Catégorie créée avec succès
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/categories/{id}:
    get:
      tags:
        - Categories
      summary: Récupérer une catégorie
      description: Récupère une catégorie spécifique par son ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID de la catégorie
      responses:
        '200':
          description: Catégorie trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Routes des entreprises
...